Move JSONObject into fuel-core.
