--- fuel-android/src/main/kotlin/com/github/kittinunf/fuel/android/core/Json.kt
@@ -1,11 +1,0 @@
-package com.github.kittinunf.fuel.android.core
-
-import org.json.JSONArray
-import org.json.JSONObject
-
-class Json(val content: String) {
-
-    fun obj(): JSONObject = JSONObject(content)
-
-    fun array(): JSONArray = JSONArray(content)
-}
